requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
PyJWT>=2.8
bcrypt>=4.0
pymongo>=4.6
dnspython>=2.6      # usually needed with pymongo + MongoDB Atlas
djongo>=1.3.6
djangorestframework-simplejwt>=5.3
